+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * BrowserSync conf improvements
 * Split legacy/ to base.scss and start modularization work
 * Add layouts/_advanced-web-interface.scss
+* Upgrade stylelint to v16
+* Add `stylelint-value-no-unknown-custom-properties` to detect undefined CSS custom properties
+* Add custom `plugin/no-unused-custom-properties` to detect unused CSS custom properties
 
 ### 4.0.0-alpha.0: 2025-12-21